+++ ootw/src/main/java/com/igrus/ootw/comment/repository/CommentRepository.java
@@ -1,6 +1,6 @@
 package com.igrus.ootw.comment.repository;
 
-import com.igrus.ootw.comment.entity.Comment;
+import com.igrus.ootw.comment.domain.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CommentRepository extends JpaRepository<Comment, Long> {}
