+++ ootw/src/main/java/com/igrus/ootw/postImage/repository/PostImageRepository.java
@@ -1,8 +1,8 @@
-package com.igrus.ootw.postImages.repository;
+package com.igrus.ootw.postImage.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.igrus.ootw.postImages.domain.PostImage;
+import com.igrus.ootw.postImage.domain.PostImage;
 
 public interface PostImageRepository extends JpaRepository<PostImage, Long> {
 }
